+++ script.module.trakt/lib/trakt/core/http.py
@@ -1,9 +1,13 @@
 from trakt.core.configuration import DEFAULT_HTTP_RETRY, DEFAULT_HTTP_MAX_RETRIES, DEFAULT_HTTP_TIMEOUT, \
     DEFAULT_HTTP_RETRY_SLEEP
 from trakt.core.context_stack import ContextStack
+from trakt.core.helpers import synchronized
 from trakt.core.request import TraktRequest
 
 from requests.adapters import HTTPAdapter
+from threading import Lock
+import calendar
+import datetime
 import logging
 import requests
 import socket
@@ -21,14 +25,31 @@ class HttpClient(object):
         self.configuration = ContextStack()
         self.session = None
 
-        self._build_session()
+        self._proxies = {}
+        self._validate_oauth_lock = Lock()
+
+        self.rebuild()
+
+    @property
+    def proxies(self):
+        if self.session and self.session.proxies:
+            return self.session.proxies
+
+        return self._proxies
+
+    @proxies.setter
+    def proxies(self, proxies):
+        if self.session:
+            self.session.proxies = proxies
+
+        self._proxies = proxies
 
     def configure(self, path=None):
         self.configuration.push(base_path=path)
 
         return self
 
-    def request(self, method, path=None, params=None, data=None, **kwargs):
+    def request(self, method, path=None, params=None, data=None, query=None, authenticated=False, **kwargs):
         # retrieve configuration
         ctx = self.configuration.pop()
 
@@ -46,22 +67,31 @@ class HttpClient(object):
         request = TraktRequest(
             self.client,
             method=method,
+
             path=path,
             params=params,
-            data=data,
 
+            data=data,
+            query=query,
+
+            authenticated=authenticated,
             **kwargs
         )
 
+        # Validate authentication details (OAuth)
+        if authenticated and not self.validate():
+            return None
+
+        # Prepare request
         prepared = request.prepare()
 
-        # retrying requests on errors >= 500
         response = None
 
         for i in range(max_retries + 1):
             if i > 0 :
                 log.warn('Retry # %s', i)
 
+            # Send request
             try:
                 response = self.session.send(prepared, timeout=timeout)
             except socket.gaierror as e:
@@ -72,8 +102,9 @@ class HttpClient(object):
 
                 log.warn('Encountered socket.gaierror (code: 8)')
 
-                response = self._build_session().send(prepared, timeout=timeout)
+                response = self.rebuild().send(prepared, timeout=timeout)
 
+            # Retry requests on errors >= 500 (when enabled)
             if not retry or response.status_code < 500:
                 break
 
@@ -91,13 +122,66 @@ class HttpClient(object):
     def delete(self, path=None, params=None, data=None, **kwargs):
         return self.request('DELETE', path, params, data, **kwargs)
 
-    def _build_session(self):
+    def rebuild(self):
         if self.session:
             log.info('Rebuilding session and connection pools...')
 
         # Build the connection pool
         self.session = requests.Session()
+        self.session.proxies = self.proxies
+
+        # Mount adapters
         self.session.mount('http://', HTTPAdapter(**self.adapter_kwargs))
         self.session.mount('https://', HTTPAdapter(**self.adapter_kwargs))
 
         return self.session
+
+    def validate(self):
+        config = self.client.configuration
+
+        # xAuth
+        if config['auth.login'] and config['auth.token']:
+            return True
+
+        # OAuth
+        if config['oauth.token']:
+            # Validate OAuth token, refresh if needed
+            return self._validate_oauth()
+
+        return False
+
+    @synchronized(lambda self: self._validate_oauth_lock)
+    def _validate_oauth(self):
+        config = self.client.configuration
+
+        if config['oauth.created_at'] is None or config['oauth.expires_in'] is None:
+            log.debug('OAuth - Missing "created_at" or "expires_in" parameter, unable to determine if token is still valid')
+            return True
+
+        current = calendar.timegm(datetime.datetime.utcnow().utctimetuple())
+        expires_at = config['oauth.created_at'] + config['oauth.expires_in'] - (48 * 60 * 60)
+
+        if current < expires_at:
+            return True
+
+        if not config['oauth.refresh']:
+            log.warn('OAuth - Unable to refresh expired token (token refreshing hasn\'t been enabled)')
+            return False
+
+        if not config['oauth.refresh_token']:
+            log.warn('OAuth - Unable to refresh expired token ("refresh_token" is parameter is missing)')
+            return False
+
+        # Refresh token
+        response = self.client['oauth'].token_refresh(config['oauth.refresh_token'], 'urn:ietf:wg:oauth:2.0:oob')
+
+        if not response:
+            log.warn('OAuth - Unable to refresh expired token (error occurred while trying to refresh the token)')
+            return False
+
+        # Update current configuration
+        config.current.oauth.from_response(response)
+
+        # Fire refresh event
+        self.client.emit('oauth.token_refreshed', response)
+        return True