+++ script.module.trakt/lib/trakt/interfaces/__init__.py
@@ -1,9 +1,13 @@
 from trakt.interfaces.base import Interface
 
 from trakt.interfaces.auth import *
+from trakt.interfaces.movies import *
 from trakt.interfaces.oauth import *
 from trakt.interfaces.scrobble import *
+from trakt.interfaces.search import *
+from trakt.interfaces.shows import *
 from trakt.interfaces.sync import *
+from trakt.interfaces.users import *
 
 INTERFACES = [
     # /
@@ -11,6 +15,7 @@ INTERFACES = [
     OAuthInterface,
 
     ScrobbleInterface,
+    SearchInterface,
 
     # /sync/
     SyncInterface,
@@ -19,7 +24,17 @@ INTERFACES = [
     SyncPlaybackInterface,
     SyncRatingsInterface,
     SyncWatchedInterface,
-    SyncWatchlistInterface
+    SyncWatchlistInterface,
+
+    # /shows/
+    ShowsInterface,
+
+    # /movies/
+    MoviesInterface,
+
+    # /users/
+    UsersInterface,
+    UsersSettingsInterface
 ]